Fix inheritance of environment variables in WindowsSubprocessFactory.
JavaSubprocessFactory causes the environment to be inherited if SubprocessBuilder#setEnv is never called or called with null. WindowsSubprocessFactory currently interprets it as an empty environment (other than the unconditionally inherited SYSTEMROOT and SYSTEMDRIVE). This CL makes the behavior on Windows the same as elsewhere.

Fixes #22190. The test failures were due to the //src/tools/remote:worker binary spawned by the test being unable to locate the runfiles directory, as the RUNFILES_MANIFEST_FILE environment variable wasn't inherited. It's unclear why the tests used to work, although I suspect it has to do with one of the many runfiles-related changes submitted to the 8.x tree.

These tests are the only uses of SubprocessBuilder without an explicit setEnv call, so other callsites are unaffected.

PiperOrigin-RevId: 631875782
Change-Id: Ifdb5055c60510bf27c72a3cd737ac2300865ce1b